integrate cli ops
