--- ops.go
@@ -1,12 +1,0 @@
-package main
-
-// Ops structure
-type Ops struct {
-	Config   string
-	Root     string
-	Protocol string
-	Addr     string
-	Delay    int
-	LogLevel string
-	Version  bool
-}
